refactor: 💡 merge DatasetError with JobRunnerError
since they can also appear in the cache. We should change the name
though
