+++ libs/libcommon/src/libcommon/dataset.py
@@ -1,133 +1,23 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright 2022 The HuggingFace Authors.
 
-from http import HTTPStatus
-from typing import Literal, Optional
+from typing import Optional
 
 import requests
 from huggingface_hub.hf_api import DatasetInfo, HfApi
 from huggingface_hub.utils._errors import RepositoryNotFoundError, RevisionNotFoundError
 from huggingface_hub.utils._headers import build_hf_headers
 
-from libcommon.exceptions import CustomError
-
-DatasetErrorCode = Literal[
-    "AskAccessHubRequestError",
-    "DatasetInfoHubRequestError",
-    "DatasetNotFoundError",
-    "DatasetRevisionNotFoundError",
-    "DisabledViewerError",
-    "GatedDisabledError",
-    "GatedExtraFieldsError",
-]
-
-
-class DatasetError(CustomError):
-    """Base class for dataset exceptions."""
-
-    def __init__(
-        self,
-        message: str,
-        status_code: HTTPStatus,
-        code: DatasetErrorCode,
-        cause: Optional[BaseException] = None,
-        disclose_cause: bool = False,
-    ):
-        super().__init__(
-            message=message, status_code=status_code, code=str(code), cause=cause, disclose_cause=disclose_cause
-        )
-
-
-class AskAccessHubRequestError(DatasetError):
-    """Raised when the request to the Hub's ask-access endpoint times out."""
-
-    def __init__(self, message: str, cause: Optional[BaseException] = None):
-        super().__init__(
-            message=message,
-            status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
-            code="AskAccessHubRequestError",
-            cause=cause,
-            disclose_cause=False,
-        )
-
-
-class DatasetInfoHubRequestError(DatasetError):
-    """Raised when the request to the Hub's dataset-info endpoint times out."""
-
-    def __init__(self, message: str, cause: Optional[BaseException] = None):
-        super().__init__(
-            message=message,
-            status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
-            code="DatasetInfoHubRequestError",
-            cause=cause,
-            disclose_cause=False,
-        )
-
-
-class DatasetNotFoundError(DatasetError):
-    """Raised when the dataset does not exist."""
-
-    def __init__(self, message: str, cause: Optional[BaseException] = None):
-        super().__init__(
-            message=message,
-            status_code=HTTPStatus.NOT_FOUND,
-            code="DatasetNotFoundError",
-            cause=cause,
-            disclose_cause=False,
-        )
-
-
-class DatasetRevisionNotFoundError(DatasetError):
-    """Raised when the dataset revision (git branch) does not exist."""
-
-    def __init__(self, message: str, cause: Optional[BaseException] = None):
-        super().__init__(
-            message=message,
-            status_code=HTTPStatus.NOT_FOUND,
-            code="DatasetRevisionNotFoundError",
-            cause=cause,
-            disclose_cause=True,
-        )
-
-
-class DisabledViewerError(DatasetError):
-    """Raised when the dataset viewer is disabled."""
-
-    def __init__(self, message: str, cause: Optional[BaseException] = None):
-        super().__init__(
-            message=message,
-            status_code=HTTPStatus.NOT_FOUND,
-            code="DisabledViewerError",
-            cause=cause,
-            disclose_cause=False,
-        )
-
-
-class GatedDisabledError(DatasetError):
-    """Raised when the dataset is gated, but disabled."""
-
-    def __init__(self, message: str, cause: Optional[BaseException] = None):
-        super().__init__(
-            message=message,
-            status_code=HTTPStatus.NOT_FOUND,
-            code="GatedDisabledError",
-            cause=cause,
-            disclose_cause=False,
-        )
-
-
-class GatedExtraFieldsError(DatasetError):
-    """Raised when the dataset is gated, with extra fields."""
-
-    def __init__(self, message: str, cause: Optional[BaseException] = None):
-        super().__init__(
-            message=message,
-            status_code=HTTPStatus.NOT_FOUND,
-            code="GatedExtraFieldsError",
-            cause=cause,
-            disclose_cause=False,
-        )
-
+from libcommon.exceptions import (
+    AskAccessHubRequestError,
+    CustomError,
+    DatasetInfoHubRequestError,
+    DatasetNotFoundError,
+    DatasetRevisionNotFoundError,
+    DisabledViewerError,
+    GatedDisabledError,
+    GatedExtraFieldsError,
+)
 
 DOES_NOT_EXIST_OR_PRIVATE_DATASET_ERROR_MESSAGE = (
     "The dataset does not exist on the Hub, or is private. Private datasets are not yet supported."
@@ -221,7 +111,7 @@ def is_supported(dataset_info: DatasetInfo) -> bool:
     """
     try:
         raise_if_not_supported(dataset_info)
-    except DatasetError:
+    except CustomError:
         return False
     return True
 
@@ -277,7 +167,7 @@ def get_dataset_info_for_supported_datasets(
             dataset_info = HfApi(endpoint=hf_endpoint).dataset_info(
                 repo_id=dataset, token=hf_token, timeout=hf_timeout_seconds
             )
-    except DatasetError as err:
+    except CustomError as err:
         raise err
     except RepositoryNotFoundError as err:
         raise DatasetNotFoundError(DOES_NOT_EXIST_OR_PRIVATE_DATASET_ERROR_MESSAGE, cause=err) from err
