+++ services/admin/src/admin/routes/dataset_backfill.py
@@ -4,7 +4,8 @@
 import logging
 from typing import Optional
 
-from libcommon.dataset import DatasetError, get_dataset_git_revision
+from libcommon.dataset import get_dataset_git_revision
+from libcommon.exceptions import JobRunnerError
 from libcommon.processing_graph import ProcessingGraph
 from libcommon.state import DatasetState
 from starlette.requests import Request
@@ -52,7 +53,7 @@ def create_dataset_backfill_endpoint(
                 {"status": "ok", "message": f"Backfilling dataset. Tasks: {tasks_list}"},
                 max_age=0,
             )
-        except (DatasetError, AdminCustomError) as e:
+        except (JobRunnerError, AdminCustomError) as e:
             return get_json_admin_error_response(e, max_age=0)
         except Exception as e:
             return get_json_admin_error_response(UnexpectedError("Unexpected error.", e), max_age=0)
